+++ extensions/epaper/core/widgets/__init__.py
@@ -2,8 +2,10 @@ from .text import TextWidget
 from .date import DateWidget
 from .roomcalendar import RoomCalendarWidget
 from .weather import WeatherNowWidget, WeatherForecastWidget, WeatherChartWidget
+from .image import ImageWidget
 
 __all__ = [
     "TextWidget", "DateWidget", "RoomCalendarWidget",
     "WeatherNowWidget", "WeatherForecastWidget", "WeatherChartWidget",
+    "ImageWidget",
 ]